moved RasterUniqueValuesLegend
