+++ src/components/controls/RasterUniqueValuesLegend.svelte
